requirements.txt
flask
python-dotenv
psycopg2-binary
Flask-CORS
bcrypt
jwt
PyJWT
Flask-JWT-Extended
flask-sqlalchemy
redis
jwt